01-naive-rag/requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
streamlit
pypdf
python-dotenv
